move check cmd tests to own module
